przerabianie metod w oparciu o security

+++ src/main/java/pl/zajavka/infrastructure/security/RoleRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends JpaRepository<RoleEntity, Long> {
+public interface RoleRepository extends JpaRepository<RoleEntity, Integer> {
 
     RoleEntity findByRole(String role);
 }
